Move CompilerDiagnostic(Builder) to a separate sub-module.
